[jira] [HBASE-5744] [89-fb] Thrift server metrics should be long instead of int
Summary: As we measure our Thrift call latencies in nanoseconds, we need to make
latencies long instead of int everywhere. There is a bug where we truncate a
nanosecond latency to int, which is a problem with RPCs that take more than
2.147483647 seconds to process.

Test Plan:
TestThriftServer is updated to test for the failure case (an RPC is artificially
made to take 3 seconds). The new test case fails without the fix.

Re-run all unit tests.

Reviewers: schen, dhruba, kannan, liyintang

Reviewed By: schen

CC: stack

Differential Revision: https://reviews.facebook.net/D2679

git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.89-fb@1310973 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/java/org/apache/hadoop/hbase/thrift/HbaseHandlerMetricsProxy.java
@@ -63,7 +63,7 @@ public class HbaseHandlerMetricsProxy implements InvocationHandler {
     try {
       long start = now();
       result = m.invoke(handler, args);
-      int processTime = (int)(now() - start);
+      long processTime = now() - start;
       metrics.incMethodTime(m.getName(), processTime);
     } catch (InvocationTargetException e) {
       throw e.getTargetException();
